api/requirements.txt
python-dotenv
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
PyMySQL==1.1.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
